create subclass `ErosionSimulationData` for the erosion module
